feat(rpc): EthApiBuilder moved to a separate builder.rs file & build_inner func added
