+++ crates/rpc/rpc/src/eth/mod.rs
@@ -1,5 +1,6 @@
 //! Sever implementation of `eth` namespace API.
 
+pub mod builder;
 pub mod bundle;
 pub mod core;
 pub mod filter;
@@ -8,8 +9,9 @@ pub mod pubsub;
 pub mod sim_bundle;
 
 /// Implementation of `eth` namespace API.
+pub use builder::EthApiBuilder;
 pub use bundle::EthBundle;
-pub use core::{EthApi, EthApiBuilder};
+pub use core::EthApi;
 pub use filter::EthFilter;
 pub use pubsub::EthPubSub;
